# src/main/resources/swagger.yaml
...
swagger: "2.0"
info:
  description: "Medikm Apis Description"
  version: "1.0"
  title: "Medikm Apis"
...
schemes:
- "http"
- "https"
paths:
...
  /patient/addPatient:
    post:
      tags:
      - "Patient Controller"
      summary: "Add Patient"
      description: "Patient Added."
      operationId: "addPatient"
      parameters:
      - in: "body"
        name: "body"
        description: "Patient Request"
        required: false
        schema:
          $ref: "#/definitions/Patient Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Patient Response"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/Patient Response"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/Patient Response"
  /patient/editPatient:
    post:
      tags:
      - "Patient Controller"
      summary: "Edit Patient"
      description: "Patient Edited."
      operationId: "editPatient"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Patient Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /tumorboardDiscussion/addNewTumorBoardDiscussion:
    post:
      tags:
      - "Tumor Board Discussion Controller"
      summary: "Add New Tumor Board Discussion"
      description: "New Tumor Board Discussion Added."
      operationId: "addNewTbDiscussion"
      parameters:
      - in: "body"
        name: "body"
        description: "Tumor Board Discussion Request"
        required: false
        schema:
          $ref: "#/definitions/Tumor Board Discussio Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Tumor Board Discussion Response"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/Tumor Board Discussion Response"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/Tumor Board Discussion Response"
  /tumorboardDiscussion/addNewTumorBoardDiscussionForum:
    post:
      tags:
      - "Tumor Board Discussion Controller"
      summary: "Add New Tumor Board Discussion Forum"
      description: "New Tumor Board Discussion Forum Added."
      operationId: "addNewTbDiscussionForum"
      parameters:
      - in: "body"
        name: "body"
        description: "Tumor Board Discussion Forum Request"
        required: false
        schema:
          $ref: "#/definitions/Tumor Board Discussion Response"
      responses:
        200:
          description: "Success, response = MedikmResponse.class"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /tumorboardDiscussion/addTumorBoardDiscussion:
    post:
      tags:
      - "Tumor Board Discussion Controller"
      summary: "Add Tumor Board Discussion"
      description: "Tumor Board Discussion Added."
      operationId: "addTumorBoardDiscussion"
      parameters:
      - in: "body"
        name: "body"
        description: "Tumor Board Discussion Request"
        required: false
        schema:
          $ref: "#/definitions/Tumor Board Discussio Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Tumor Board Discussion Response"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/Tumor Board Discussion Response"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/Tumor Board Discussion Response"
  /tumorboardDiscussion/addTumorBoardDiscussionForum:
    post:
      tags:
      - "Tumor Board Discussion Controller"
      summary: "Add Tumor Board Discussion Forum"
      description: "Tumor Board Discussion Forum Added."
      operationId: "addTbDiscussionForum"
      parameters:
      - in: "body"
        name: "body"
        description: "Tumor Board Discussion Forum Request"
        required: false
        schema:
          $ref: "#/definitions/Tumor Board Discussion Response"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
definitions:
...
  Case1:
    type: "object"
    properties:
      caseId:
        type: "integer"
        format: "int32"
      age:
        type: "integer"
        format: "int32"
      tumorSize:
        type: "number"
        format: "double"
      noNodes:
        type: "integer"
        format: "int32"
      distantMets:
        type: "integer"
        format: "int32"
      distantMetsComm:
        type: "string"
      clinicalStage:
        type: "string"
      histologicalGrade:
        type: "integer"
        format: "int32"
      geteRStatus:
        type: "string"
      geteRPercent:
        type: "integer"
        format: "int32"
      getpRStatus:
        type: "string"
      getpRPercent:
        type: "integer"
        format: "int32"
      caseHER2neuAmplified:
        type: "integer"
        format: "int32"
      gethER2Status:
        type: "string"
      her2Sish:
        type: "string"
      primaryDiagnosisDt:
        type: "string"
        format: "date-time"
      deceasedDueBC:
        type: "integer"
        format: "int32"
      treatmentRelatedDeath:
        type: "integer"
        format: "int32"
      unrelatedDeath:
        type: "integer"
        format: "int32"
      dtDeath:
        type: "string"
        format: "date-time"
      dateFirstEncounter:
        type: "string"
        format: "date-time"
      menopauseStatus:
        type: "string"
      significanceRisk:
        type: "boolean"
        default: false
      significanceComorbities:
        type: "boolean"
        default: false
      version:
        type: "integer"
        format: "int32"
      clinicalStagingMethod:
        type: "integer"
        format: "int32"
      getwHOClassification:
        type: "integer"
        format: "int32"
      pleuralInv:
        type: "integer"
        format: "int32"
      necrosisPercent:
        type: "integer"
        format: "int32"
      necrosisPresent:
        type: "string"
      reasonNotTreated:
        type: "integer"
        format: "int32"
      diseaseTreated:
        type: "integer"
        format: "int32"
      seenOnCT:
        type: "integer"
        format: "int32"
      durationOfSymptomsMon:
        type: "integer"
        format: "int32"
      seenOnMRI:
        type: "integer"
        format: "int32"
      seenOnPAT:
        type: "integer"
        format: "int32"
      seenOnThalliumScan:
        type: "integer"
        format: "int32"
      seenOnXray:
        type: "integer"
        format: "int32"
      seenOnBonescan:
        type: "integer"
        format: "int32"
      tumorDepth:
        type: "integer"
        format: "int32"
      invasionOfAdjStructure:
        type: "integer"
        format: "int32"
      invasionOfPleura:
        type: "integer"
        format: "int32"
      invasionOfChestWall:
        type: "integer"
        format: "int32"
      invasionOfDiaphragm:
        type: "integer"
        format: "int32"
      invasionOfPhrenicNerve:
        type: "integer"
        format: "int32"
      invasionOfPericardium:
        type: "integer"
        format: "int32"
      atelectasis:
        type: "integer"
        format: "int32"
      obstractivePneumonitis:
        type: "integer"
        format: "int32"
      separateTNodeInSameLobe:
        type: "integer"
        format: "int32"
      invasionInMediastnum:
        type: "integer"
        format: "int32"
      invasionIntoHeart:
        type: "integer"
        format: "int32"
      invasionOfMainBronchus:
        type: "integer"
        format: "int32"
      invasionIntoGreatVessels:
        type: "integer"
        format: "int32"
      invasionIntoTrachea:
        type: "integer"
        format: "int32"
      invasionIntoRecLaryngealNerve:
        type: "integer"
        format: "int32"
      invasionIntoEsophagus:
        type: "integer"
        format: "int32"
      invasionIntoVertbralBody:
        type: "integer"
        format: "int32"
      invasionIntoCarina:
        type: "integer"
        format: "int32"
      gettNodesInDiffLobe:
        type: "integer"
        format: "int32"
      nodalMetastases:
        type: "string"
      esophagectomyDone:
        type: "integer"
        format: "int32"
      residualTumor:
        type: "integer"
        format: "int32"
      debulkingStatus:
        type: "integer"
        format: "int32"
      platinStatus:
        type: "integer"
        format: "int32"
      recurrence:
        type: "integer"
        format: "int32"
      dateRecurrence:
        type: "string"
        format: "date-time"
      secondaryDebulkingDate:
        type: "string"
        format: "date-time"
      intervalDebulkingDt:
        type: "string"
        format: "date-time"
      baseLineCA125:
        type: "string"
      prgFreeSurvival:
        type: "integer"
        format: "int32"
      overallSurvival:
        type: "integer"
        format: "int32"
      esophagealCStagingMethod:
        type: "integer"
        format: "int32"
      later:
        type: "string"
      stPrimary:
        type: "integer"
        format: "int32"
      ndPrimary:
        type: "integer"
        format: "int32"
      diseasePropagation:
        type: "integer"
        format: "int32"
      metastatic:
        type: "integer"
        format: "int32"
      chiefComplaint:
        type: "string"
      lymfovascularInvasion:
        type: "integer"
        format: "int32"
      margins:
        type: "string"
      closestMargins:
        type: "string"
      immuneHistochemistry:
        type: "string"
      noFossi:
        type: "integer"
        format: "int32"
      brca:
        type: "integer"
        format: "int32"
      oncDxScore:
        type: "integer"
        format: "int32"
      bladderNeck:
        type: "integer"
        format: "int32"
      neoadjResponse:
        type: "string"
      metastasis:
        type: "integer"
        format: "int32"
      metPrimarySite:
        type: "string"
      otherBiomarkers:
        type: "string"
      nln:
        type: "integer"
        format: "int32"
      tlocation:
        type: "string"
      caseStatus:
        type: "string"
      tumorType:
        type: "string"
      tisBankPnr:
        type: "integer"
        format: "int32"
      tisBankOdunsi:
        type: "integer"
        format: "int32"
      nvInf:
        type: "integer"
        format: "int32"
      nyso:
        type: "integer"
        format: "int32"
      wt:
        type: "integer"
        format: "int32"
      ido:
        type: "integer"
        format: "int32"
      tisBank:
        type: "integer"
        format: "int32"
      mucosalInvasion:
        type: "integer"
        format: "int32"
      mi:
        type: "string"
      lesionDepth:
        type: "string"
      dist:
        type: "string"
      trupture:
        type: "integer"
        format: "int32"
      tumorOnSurface:
        type: "integer"
        format: "int32"
      lsvi:
        type: "integer"
        format: "int32"
      parametrialInvasion:
        type: "integer"
        format: "int32"
      stromalInvasion:
        type: "integer"
        format: "int32"
      depth:
        type: "integer"
        format: "int32"
      horizontalSpread:
        type: "integer"
        format: "int32"
      washings:
        type: "integer"
        format: "int32"
      capsuleIntact:
        type: "integer"
        format: "int32"
      surgicalSpill:
        type: "integer"
        format: "int32"
      metastaticSite:
        type: "string"
      lnMetastasisSite:
        type: "string"
      malignantAscites:
        type: "integer"
        format: "int32"
      grossCervicalInvolvement:
        type: "integer"
        format: "int32"
      aorticNodePositive:
        type: "integer"
        format: "int32"
      tumorPenetratedSerosalSurface:
        type: "integer"
        format: "int32"
      malignantCellsInAscites:
        type: "integer"
        format: "int32"
      malignantCellsInPeritonealWashings:
        type: "integer"
        format: "int32"
      peritonealImplants:
        type: "integer"
        format: "int32"
      pelvicExtensions:
        type: "integer"
        format: "int32"
      extensionToOtherStructures:
        type: "integer"
        format: "int32"
      specify:
        type: "string"
      operable:
        type: "integer"
        format: "int32"
      fertilitySparing:
        type: "integer"
        format: "int32"
      hydronephrosis:
        type: "integer"
        format: "int32"
      clinicallyVisibleLesion:
        type: "integer"
        format: "int32"
      surgicalCandidate:
        type: "integer"
        format: "int32"
      surgicallyResectable:
        type: "integer"
        format: "int32"
      residualDisease:
        type: "integer"
        format: "int32"
      malignantEpithelialCarcinoma:
        type: "integer"
        format: "int32"
      pureEndometrioidCarcinoma:
        type: "integer"
        format: "int32"
      serousCellCarcinoma:
        type: "integer"
        format: "int32"
      stromalTumor:
        type: "integer"
        format: "int32"
      cap:
        type: "string"
      uterusDiseaseLimited:
        type: "integer"
        format: "int32"
      suspectedExtrauterineDisease:
        type: "integer"
        format: "int32"
      myometrialInvasion:
        type: "integer"
        format: "int32"
      adverseRiskFactors:
        type: "integer"
        format: "int32"
      details:
        type: "string"
      clearCellHistology:
        type: "integer"
        format: "int32"
      numberOfMetastaticSites:
        type: "integer"
        format: "int32"
      miscellaneous:
        type: "string"
      tumorVolume:
        type: "string"
      cancerStatus:
        type: "string"
      progressionFreeSurvivalAfterSurgery:
        type: "string"
      progressionFreeSurvivalAfterRt:
        type: "string"
      progressionFreeSurvivalAfterBt:
        type: "string"
      alfhaFp:
        type: "integer"
        format: "int32"
      betaHcg:
        type: "integer"
        format: "int32"
      ldh:
        type: "integer"
        format: "int32"
      prostateVolume:
        type: "integer"
        format: "int32"
      biopsyGleasonScore:
        type: "string"
      surgicalGleasonScore:
        type: "string"
      numberPositiveBiopsyCores:
        type: "integer"
        format: "int32"
      positiveBiopsyCoresLocation:
        type: "string"
      psaDoublingTime:
        type: "string"
      psaDensity:
        type: "number"
        format: "double"
      psaCureRate:
        type: "number"
        format: "float"
      psaLevels:
        type: "number"
        format: "double"
      newPsaLevel:
        type: "string"
      numberOfBxCores:
        type: "integer"
        format: "int32"
      numberOfPositiveCores:
        type: "integer"
        format: "int32"
      percentOfPositiveCores:
        type: "number"
        format: "double"
      prostateEchotexture:
        type: "string"
      percantCancerInAnyCore:
        type: "number"
        format: "float"
      netLifeexpectancy:
        type: "integer"
        format: "int32"
      probabilityofCadeathse:
        type: "string"
      tnmStagingSystem:
        type: "string"
      probabilityofCadeathdt:
        type: "string"
      oneY:
        type: "number"
        format: "float"
      twoY:
        type: "number"
        format: "float"
      crpcMedianSurvivalInYears:
        type: "integer"
        format: "int32"
      probabilityOfLymphNodeInvolvement:
        type: "number"
        format: "float"
      progressionFreeSurvivalProbabilityAfterSalvageRtSixY:
        type: "string"
      extracapsularExtension:
        type: "integer"
        format: "int32"
      fixation:
        type: "integer"
        format: "int32"
      seminalVesicleInvasion:
        type: "integer"
        format: "int32"
      tumorInCentralLocation:
        type: "integer"
        format: "int32"
      indolentCancer:
        type: "number"
        format: "float"
      leftBaseMedial:
        type: "string"
      leftBaseLateral:
        type: "string"
      rightBaseMedial:
        type: "string"
      rightBaseLateral:
        type: "string"
      leftMidMedial:
        type: "string"
      leftMidLateral:
        type: "string"
      rightMidMedial:
        type: "string"
      rightMidLateral:
        type: "string"
      leftApexMedial:
        type: "string"
      leftApexLateral:
        type: "string"
      rightApexMedial:
        type: "string"
      rightApexLateral:
        type: "string"
      organConfinedDisease:
        type: "number"
        format: "float"
      capsularPenetration:
        type: "number"
        format: "float"
      lymphNodeMets:
        type: "number"
        format: "float"
      seminalVInvasionPercentage:
        type: "number"
        format: "float"
      tumorLocation:
        $ref: "#/definitions/TumorLocation"
      geneticRiskEvaluation:
        type: "string"
      refPhysician:
        type: "string"
      optimalcytred:
        type: "integer"
        format: "int32"
      parametriumPos:
        type: "integer"
        format: "int32"
      encounterCollection:
        type: "array"
        items:
          $ref: "#/definitions/Encounter"
      surgeryCollection:
        type: "array"
        items:
          $ref: "#/definitions/Surgery"
      caseResourcesCollection:
        type: "array"
        items:
          $ref: "#/definitions/CaseResources"
      casePhysicianCollection:
        type: "array"
        items:
          $ref: "#/definitions/CasePhysician"
      diagnosticWorkupCollection:
        type: "array"
        items:
          $ref: "#/definitions/DiagnosticWorkup"
      patient:
        $ref: "#/definitions/Patient"
      diseaseType:
        $ref: "#/definitions/DiseaseType"
      diseaseSubTypeId:
        $ref: "#/definitions/DiseaseSubType"
      careFacility:
        $ref: "#/definitions/CareFacility"
      functionalStatus:
        $ref: "#/definitions/Functionalstatus"
      genetics:
        $ref: "#/definitions/Genetics"
      genomics:
        $ref: "#/definitions/Gemonics"
      finalTNMStageId:
        $ref: "#/definitions/TnmStage"
      workingTNMStageId:
        $ref: "#/definitions/TnmStage"
      referingPhysiciaId:
        $ref: "#/definitions/Physician"
      medOncPhysicianId:
        $ref: "#/definitions/Physician"
      radOncPhysicianId:
        $ref: "#/definitions/Physician"
      plasticSurgeonId:
        $ref: "#/definitions/Physician"
      surgOncPhysicianId:
        $ref: "#/definitions/Physician"
      pathologist:
        $ref: "#/definitions/Physician"
      gynOncologist:
        $ref: "#/definitions/Physician"
      cancerSubTypeId:
        $ref: "#/definitions/CancerSubType"
      cancerTypeId:
        $ref: "#/definitions/CancerType"
      biopsyCollection:
        type: "array"
        items:
          $ref: "#/definitions/Biopsy"
      tumorBoardDiscussionCollection:
        type: "array"
        items:
          $ref: "#/definitions/TumorBoardDiscussion"
      chemotherapyCollection:
        type: "array"
        items:
          $ref: "#/definitions/Chemotherapy"
      episodeCollection:
        type: "array"
        items:
          $ref: "#/definitions/Episode"
      pathologyCollection:
        type: "array"
        items:
          $ref: "#/definitions/Pathology"
      physicalExamCollection:
        type: "array"
        items:
          $ref: "#/definitions/PhysicalExam"
      radiationTherapyCollection:
        type: "array"
        items:
          $ref: "#/definitions/RadiationTherapy"
      cancerStageCollection:
        type: "array"
        items:
          $ref: "#/definitions/CancerStage"
      forumAnswereResourcesCollection:
        type: "array"
        items:
          $ref: "#/definitions/ForumAnswereResources"
      followupCollection:
        type: "array"
        items:
          $ref: "#/definitions/Followup"
      metPrimary:
        type: "integer"
        format: "int32"
      tumorBoardDiscussionQuestionList:
        type: "array"
        items:
          $ref: "#/definitions/TumorBoardDiscussionQuestion"
      diseaseSymptomsCollection:
        type: "array"
        items:
          $ref: "#/definitions/DiseaseSymptoms"
      caseQualityIndicatorCollection:
        type: "array"
        items:
          $ref: "#/definitions/CaseQualityIndicator"
      createdBy:
        $ref: "#/definitions/Physician"
      lastUpdatedBy:
        $ref: "#/definitions/Physician"
      creationTime:
        type: "string"
        format: "date-time"
      pathologicalT:
        type: "string"
      erstatus:
        type: "string"
      erpercent:
        type: "integer"
        format: "int32"
      prstatus:
        type: "string"
      prpercent:
        type: "integer"
        format: "int32"
      her2Status:
        type: "string"
      whoclassification:
        type: "integer"
        format: "int32"
      tnodesInDiffLobe:
        type: "integer"
        format: "int32"
...
  FamilyHistory:
    type: "object"
    properties:
      familyHistoryPK:
        $ref: "#/definitions/FamilyHistoryPK"
      noFirstDegreeRelativeswithBc:
        type: "integer"
        format: "int32"
      youngestAgeDiagnosisFirstDegreeRelatives:
        type: "integer"
        format: "int32"
      anyRelativeBcBefore50:
        type: "integer"
        format: "int32"
      ovarianCancer:
        type: "integer"
        format: "int32"
      ovarianandBc:
        type: "integer"
        format: "int32"
      maleRelativewithBc:
        type: "integer"
        format: "int32"
      bilateralBc:
        type: "integer"
        format: "int32"
      geneticAbnormalities:
        type: "integer"
        format: "int32"
      geneticAbnormalitiesComm:
        type: "string"
      noFirstDegreeRelativesBC:
        type: "integer"
        format: "int32"
      brca12:
        type: "integer"
        format: "int32"
      otherCancer:
        type: "string"
      thyroidDisease:
        type: "integer"
        format: "int32"
      familyHistoryComments:
        type: "string"
      patient:
        $ref: "#/definitions/Patient"
...
  Functionalstatus:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      status:
        type: "string"
      patientCollection:
        type: "array"
        items:
          $ref: "#/definitions/Patient"
...
  MedikmResponse:
    type: "object"
    properties:
      responseCode:
        type: "string"
        description: "Response Code"
      responseMessage:
        type: "string"
        description: "Response Message"
      authenticationKey:
        type: "string"
        description: "Authentication Key"
    description: "MedikmResponse class used for returning medikmresponse"
...
  Patient:
    type: "object"
    properties:
      patientId:
        type: "integer"
        format: "int32"
      phoneNumber:
        type: "string"
      emergencyContactName:
        type: "string"
      emergencyContactPhoneNumber:
        type: "string"
      zipcode:
        type: "string"
      birthDate:
        type: "string"
        format: "date-time"
      increaseRisk:
        type: "integer"
        format: "int32"
      gender:
        type: "string"
      patientConcerns:
        type: "string"
      cellPhoneNumber:
        type: "string"
      workPhoneNumber:
        type: "string"
      address:
        type: "string"
      city:
        type: "string"
      country:
        type: "string"
      employementType:
        type: "string"
      insuranceName:
        type: "string"
      mammographicBreastDensity:
        type: "string"
      status:
        type: "string"
      active:
        type: "integer"
        format: "int32"
      medRecNo:
        type: "integer"
        format: "int32"
      ethnicity:
        type: "integer"
        format: "int32"
      case1Collection:
        type: "array"
        items:
          $ref: "#/definitions/Case1"
      reproductiveHistoryCollection:
        type: "array"
        items:
          $ref: "#/definitions/ReproductiveHistory"
      personalMedicalHistoryCollection:
        type: "array"
        items:
          $ref: "#/definitions/PersonalMedicalHistory"
      functionalStatus:
        $ref: "#/definitions/Functionalstatus"
      race:
        $ref: "#/definitions/Race"
      relativeDetailsCollection:
        type: "array"
        items:
          $ref: "#/definitions/RelativeDetails"
      familyHistoryCollection:
        type: "array"
        items:
          $ref: "#/definitions/FamilyHistory"
      mname:
        type: "string"
      lname:
        type: "string"
      fname:
        type: "string"
...
  Race:
    type: "object"
    properties:
      typeId:
        type: "integer"
        format: "int32"
      race:
        type: "string"
      patientCollection:
        type: "array"
        items:
          $ref: "#/definitions/Patient"
...
  RelativeDetails:
    type: "object"
    properties:
      relativeDetailsPK:
        $ref: "#/definitions/RelativeDetailsPK"
      ageatDiagnosis:
        type: "integer"
        format: "int32"
      ageatDeath:
        type: "integer"
        format: "int32"
      patient:
        $ref: "#/definitions/Patient"
      relationship:
        $ref: "#/definitions/Relationship"
      genetics:
        $ref: "#/definitions/Genetics"
      gemonics:
        $ref: "#/definitions/Gemonics"
      cancerSubType:
        $ref: "#/definitions/CancerSubType"
      cancerType:
        $ref: "#/definitions/CancerType"
...
  ReproductiveHistory:
    type: "object"
    properties:
      reproductiveHistoryPK:
        $ref: "#/definitions/ReproductiveHistoryPK"
      ageMenarche:
        type: "integer"
        format: "int32"
      ageMenopause:
        type: "integer"
        format: "int32"
      lastMenstrualDt:
        type: "string"
        format: "date-time"
      noPregnancies:
        type: "integer"
        format: "int32"
      noDeliveries:
        type: "integer"
        format: "int32"
      ageFirstLiveBirth:
        type: "integer"
        format: "int32"
      breastFeedingFor6Months:
        type: "integer"
        format: "int32"
      ageFirstBirth:
        type: "integer"
        format: "int32"
      progesteroneTherapy:
        type: "integer"
        format: "int32"
      progesteroneTherapyComm:
        type: "string"
      oralContracceptives:
        type: "integer"
        format: "int32"
      oralContracceptivesComm:
        type: "string"
      fertilityDrugs:
        type: "integer"
        format: "int32"
      fertilityDrugsComm:
        type: "string"
      birthControlPillsUsed:
        type: "boolean"
        default: false
      patient:
        $ref: "#/definitions/Patient"
...
  TumorLocation:
    type: "object"
    properties:
      tumorLocationId:
        type: "integer"
        format: "int32"
      tumorLocationName:
        type: "string"
      case1Collection:
        type: "array"
        items:
          $ref: "#/definitions/Case1"
      diseaseTypeId:
        $ref: "#/definitions/DiseaseType"
      diseaseSubTypeId:
        $ref: "#/definitions/DiseaseSubType"
...